pyproject.toml
[project]
name = "upworthy-autosearch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pandas", "scikit-learn", "requests", "openai", "google-generativeai", "tqdm", "numpy", "pyarrow", "jinja2"]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/upworthy_autosearch"]
